u16

+++ src/lib.rs
@@ -76,13 +76,13 @@ impl ErrorCode {
 pub const DEFAULT_ARENA_SIZE: usize = 10000;
 
 #[derive(Clone, Copy, PartialEq, Eq, Debug)]
-pub struct ArenaRef(pub u32);
+pub struct ArenaRef(pub u16);
 
 impl ArenaRef {
-    pub const NULL: ArenaRef = ArenaRef(u32::MAX);
+    pub const NULL: ArenaRef = ArenaRef(u16::MAX);
 
     pub fn is_null(self) -> bool {
-        self.0 == u32::MAX
+        self.0 == u16::MAX
     }
 }
 
@@ -102,7 +102,7 @@ pub enum Value {
 #[derive(Debug)]
 pub struct Arena<const N: usize> {
     pub values: [Cell<Value>; N],
-    pub refcounts: [Cell<u32>; N],
+    pub refcounts: [Cell<u16>; N],
     pub next_free: Cell<usize>,
 }
 
@@ -124,7 +124,7 @@ impl<const N: usize> Arena<N> {
                 self.values[idx].set(value);
                 self.refcounts[idx].set(1);
                 self.next_free.set((idx + 1) % N);
-                return Ok(ArenaRef(idx as u32));
+                return Ok(ArenaRef(idx as u16));
             }
         }
 
@@ -150,7 +150,7 @@ impl<const N: usize> Arena<N> {
         let idx = r.0 as usize;
         if !matches!(self.values[idx].get(), Value::Free) {
             let rc = self.refcounts[idx].get();
-            if rc < u32::MAX {
+            if rc < u16::MAX {
                 self.refcounts[idx].set(rc + 1);
             }
         }
